requirements.txt
pyyaml
pypdf
python-docx
openpyxl
requests
